more modern React Email usage
