--- with-express-react-email/transactional/package.json
@@ -1,13 +1,0 @@
-{
-  "name": "emails",
-  "version": "0.0.0",
-  "private": true,
-  "scripts": {
-    "dev": "email dev",
-    "export": "email export"
-  },
-  "dependencies": {
-    "@react-email/components": "latest",
-    "react-email": "latest"
-  }
-}
